ml-model/requirements.txt
# Core ML libraries
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.15.0
transformers>=4.35.0

# Computer Vision
opencv-python>=4.8.1
Pillow>=10.1.0
albumentations>=1.3.1

# Data processing
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Model serving
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Image processing
imageio>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
datasets>=2.14.0
huggingface-hub>=0.17.0

# Development
jupyter>=1.0.0
notebook>=7.0.0
